Fixes ServiceParser not implementing the 'share' option correctly.
The ServiceParser sometimes checks for 'shared' set to true, rather than the documented 'share'. This option determines if the service is registered with the Pimple $app->share() method, which causes the service instance to be re-used across service-locator requests.

The option is documented as 'share': https://studionone.github.io/flint/services.html#syntax-for-app-services-php

This is particularly bad as:

 - we check for 'shared' in ServiceParser::parse(), and only use $app->share() if 'shared' is set to true. This means services with 'share' set to true aren't registered that way, against documentation.

 - ServiceParser::loadControllers() still checks and sets 'share', but relies on ServiceParser::parse() to apply the behaviour. This means that none of our controller services are being registered as shared, despite this being the desired behaviour.